improve order process

+++ inventory-service/src/main/kotlin/com/siberia/market/inventory/model/Stock.kt
@@ -31,5 +31,5 @@ data class Stock(
     val item: Item,
 
     @Column(name = "item_count")
-    val itemCount: Long
+    val itemCount: Int
 )
